+++ native_library/android/AdMobANE/app/src/main/java/com/tuarua/admobane/RewardedVideoController.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017 Tua Rua Ltd.
+ *  Copyright 2018 Tua Rua Ltd.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -17,7 +17,9 @@
 
 package com.tuarua.admobane
 
+import android.os.Bundle
 import com.adobe.fre.FREContext
+import com.google.ads.mediation.admob.AdMobAdapter
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.MobileAds
 import com.google.android.gms.ads.reward.RewardItem
@@ -27,7 +29,8 @@ import com.google.gson.Gson
 import com.tuarua.admobane.Position.*
 import com.tuarua.frekotlin.FreKotlinController
 
-class RewardedVideoController(override var context: FREContext?) : FreKotlinController, RewardedVideoAdListener {
+class RewardedVideoController(override var context: FREContext?,
+                              private val isPersonalised: Boolean) : FreKotlinController, RewardedVideoAdListener {
     private val gson = Gson()
     private var _adView: RewardedVideoAd? = null
     private var _showOnLoad: Boolean = true
@@ -48,11 +51,12 @@ class RewardedVideoController(override var context: FREContext?) : FreKotlinCont
         av.rewardedVideoAdListener = this
 
         val builder = AdRequest.Builder()
+        if (!isPersonalised) {
+            val extras = Bundle()
+            extras.putString("npa", "1")
+            builder.addNetworkExtrasBundle(AdMobAdapter::class.java, extras)
+        }
         if (targeting != null) {
-            builder.setGender(targeting.gender)
-            if (targeting.birthday != null) {
-                builder.setBirthday(targeting.birthday)
-            }
             if (targeting.forChildren != null) {
                 val forChildren = targeting.forChildren
                 forChildren?.let { builder.tagForChildDirectedTreatment(it) }
@@ -70,11 +74,11 @@ class RewardedVideoController(override var context: FREContext?) : FreKotlinCont
     }
 
     override fun onRewardedVideoAdClosed() {
-        sendEvent(Constants.ON_CLOSED, gson.toJson(AdMobEvent(REWARD.ordinal)))
+        dispatchEvent(Constants.ON_CLOSED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdLeftApplication() {
-        sendEvent(Constants.ON_LEFT_APPLICATION, gson.toJson(AdMobEvent(REWARD.ordinal)))
+        dispatchEvent(Constants.ON_LEFT_APPLICATION, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdLoaded() {
@@ -82,28 +86,32 @@ class RewardedVideoController(override var context: FREContext?) : FreKotlinCont
         if (_showOnLoad) {
             av.show()
         }
-        sendEvent(Constants.ON_LOADED, gson.toJson(AdMobEvent(REWARD.ordinal)))
+        dispatchEvent(Constants.ON_LOADED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdOpened() {
-        sendEvent(Constants.ON_OPENED, gson.toJson(AdMobEvent(REWARD.ordinal)))
+        dispatchEvent(Constants.ON_OPENED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewarded(rewardItem: RewardItem?) {
         if (rewardItem != null) {
-            sendEvent(Constants.ON_REWARDED,
+            dispatchEvent(Constants.ON_REWARDED,
                     gson.toJson(AdMobEventWithReward(REWARD.ordinal, amount = rewardItem.amount, type = rewardItem.type)))
         } else {
-            sendEvent(Constants.ON_REWARDED, gson.toJson(AdMobEventWithReward(REWARD.ordinal)))
+            dispatchEvent(Constants.ON_REWARDED, gson.toJson(AdMobEventWithReward(REWARD.ordinal)))
         }
     }
 
+    override fun onRewardedVideoCompleted() {
+        dispatchEvent(Constants.ON_VIDEO_COMPLETE, gson.toJson(AdMobEvent(REWARD.ordinal)))
+    }
+
     override fun onRewardedVideoStarted() {
-        sendEvent(Constants.ON_VIDEO_STARTED, gson.toJson(AdMobEvent(REWARD.ordinal)))
+        dispatchEvent(Constants.ON_VIDEO_STARTED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdFailedToLoad(p0: Int) {
-        sendEvent(Constants.ON_LOAD_FAILED, gson.toJson(AdMobEvent(REWARD.ordinal, p0)))
+        dispatchEvent(Constants.ON_LOAD_FAILED, gson.toJson(AdMobEvent(REWARD.ordinal, p0)))
     }
 
     override val TAG: String