+++ native_library/ios/AdMobANE/AdMobANE_FW/RewardVideoController.swift
@@ -23,12 +23,13 @@ class RewardVideoController: NSObject, FreSwiftController, GADRewardBasedVideoAd
     internal var context: FreContextSwift!
     private var _showOnLoad: Bool = true
     private var _airVC: UIViewController?
+    private var isPersonalised: Bool = true
     private var adView: GADRewardBasedVideoAd?
     
-    convenience init(context: FreContextSwift) {
+    convenience init(context: FreContextSwift, isPersonalised: Bool) {
         self.init()
         self.context = context
-        
+        self.isPersonalised = isPersonalised
     }
     func load(airVC: UIViewController, unitId: String, deviceList: [String]?,
               targeting: Targeting?, showOnLoad: Bool) {
@@ -38,22 +39,22 @@ class RewardVideoController: NSObject, FreSwiftController, GADRewardBasedVideoAd
         
         adView?.delegate = self
         let request = GADRequest()
+        if !isPersonalised {
+            let extras = GADExtras()
+            extras.additionalParameters = ["npa": "1"]
+            request.register(extras)
+        }
         if deviceList != nil {
             request.testDevices = deviceList!
         }
-        
+
         if let t = targeting {
             if let fc = t.forChildren {
                 request.tag(forChildDirectedTreatment: fc)
             }
-            request.gender = t.gender
-            if let birthday = t.birthday {
-                request.birthday = birthday
-            }
             if let contentUrl = t.contentUrl {
                 request.contentURL = contentUrl
             }
-            
         }
         adView?.load(request, withAdUnitID: unitId)
         
@@ -74,8 +75,7 @@ class RewardVideoController: NSObject, FreSwiftController, GADRewardBasedVideoAd
         props["position"] = Position.reward.rawValue
         props["type"] = reward.type
         props["amount"] = reward.amount
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_REWARDED, value: json.description)
+        dispatchEvent(name: Constants.ON_REWARDED, value: JSON(props).description)
     }
     
     func dispose() {
@@ -89,8 +89,7 @@ class RewardVideoController: NSObject, FreSwiftController, GADRewardBasedVideoAd
     func rewardBasedVideoAdDidReceive(_ rewardBasedVideoAd: GADRewardBasedVideoAd) {
         var props: [String: Any] = Dictionary()
         props["position"] = Position.reward.rawValue
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_LOADED, value: json.description)
+        dispatchEvent(name: Constants.ON_LOADED, value: JSON(props).description)
         
         guard let av = adView, let avc = _airVC else {return}
         if _showOnLoad {
@@ -101,29 +100,31 @@ class RewardVideoController: NSObject, FreSwiftController, GADRewardBasedVideoAd
     func rewardBasedVideoAdDidOpen(_ rewardBasedVideoAd: GADRewardBasedVideoAd) {
         var props: [String: Any] = Dictionary()
         props["position"] = Position.reward.rawValue
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_OPENED, value: json.description)
+        dispatchEvent(name: Constants.ON_OPENED, value: JSON(props).description)
     }
     
     func rewardBasedVideoAdDidStartPlaying(_ rewardBasedVideoAd: GADRewardBasedVideoAd) {
         var props: [String: Any] = Dictionary()
         props["position"] = Position.reward.rawValue
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_VIDEO_STARTED, value: json.description)
+        dispatchEvent(name: Constants.ON_VIDEO_STARTED, value: JSON(props).description)
+    }
+    
+    func rewardBasedVideoAdDidCompletePlaying(_ rewardBasedVideoAd: GADRewardBasedVideoAd) {
+        var props: [String: Any] = Dictionary()
+        props["position"] = Position.reward.rawValue
+        dispatchEvent(name: Constants.ON_VIDEO_COMPLETE, value: JSON(props).description)
     }
     
     func rewardBasedVideoAdDidClose(_ rewardBasedVideoAd: GADRewardBasedVideoAd) {
         var props: [String: Any] = Dictionary()
         props["position"] = Position.reward.rawValue
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_CLOSED, value: json.description)
+        dispatchEvent(name: Constants.ON_CLOSED, value: JSON(props).description)
     }
     
     func rewardBasedVideoAdWillLeaveApplication(_ rewardBasedVideoAd: GADRewardBasedVideoAd) {
         var props: [String: Any] = Dictionary()
         props["position"] = Position.reward.rawValue
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_LEFT_APPLICATION, value: json.description)
+        dispatchEvent(name: Constants.ON_LEFT_APPLICATION, value: JSON(props).description)
     }
     
     func rewardBasedVideoAd(_ rewardBasedVideoAd: GADRewardBasedVideoAd,
@@ -131,8 +132,7 @@ class RewardVideoController: NSObject, FreSwiftController, GADRewardBasedVideoAd
         var props: [String: Any] = Dictionary()
         props["position"] = Position.reward.rawValue
         props["errorCode"] = 0
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_LOAD_FAILED, value: json.description)
+        dispatchEvent(name: Constants.ON_LOAD_FAILED, value: JSON(props).description)
     }
     
 }